requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pytest
spyne
lxml
PyQt5
black
python-dotenv